Close Issue #11 Created ImportedListHandler to take the imported list froma  document and turn it into a ListItem which is sent to AddToAvailable.
Created AddToAvailable and AddToSelected functions for consistancy.
These functions handle adding information to their respective WordSets and Checked Box Lists on the form.

Tidied some extra comments out of ListEditor.cs.

Renamed the button user uses for importing a word list from a text file. it is now called btnImportList

Issue #12
Rewrote WordSet Class and greatly expanded on it. WordSets are really lists of ListItems with extra functions.

Added a Count function to count number of lists in a WordSet.

Added an Exists function to check if a list exists in a WordSet.

Added an AddList function for directly adding a ListItem to a WordSet.

Added an AddWords function for adding a string and a list of words to a WordSet by converting them to a ListItem then calling AddList.

Added two RemoveLIst  functions for removing a wordlist from a WordSet.
    The base RemoveList requires a string for the name of the list to be removed. It uses Exists to check if that list exists in the WordSet.
   The second RemoveList takes a ListItem, finds the DisplayName of the ListItem then calls the base RemoveList.